Reorganize `log` module
